Application Simulation: Register Users only the lack of mechanisms to collect threads' return value.
createMultiOrgs.sh : call createOrg.sh using for loop
Client.java : increase the number of seconds for waiting for send
Proposal. Because lauch a container is not quick.
ClientRegisterUser : a script for that a client registe user. Used by
MultiClientsRegisterUser.
MultiClientsRegisterUser.java : the simulation that simulate multi
clients register users

add chaincode Status can be installed in the scripts.